Add dark scrolling style to body textarea
This change introduces a "scroll-dark" class that can be used on any
overflow-scrolling elements on dark backgrounds. The default browser
styling for scrollbars looks really out of place on a dark background,
and this change helps fix that.

From what I can tell, support for scrollbar styling via CSS varies
quite a bit from browser to browser. I tested this on Chrome, Firefox,
and Safari, and all of those browsers look reasonable, but I'm sure
there are others / different versions that won't benefit from this
change.