Handle URL errors in a more forgiving fashion
This change updates the backend server to try parsing the provided URL
before making a request with it. This helps in two ways:

  1. It gives us an opportunity to catch URL errors separate from other
     request errors, giving more granular error messages.

  2. It allows us to modify the provided URL to fix common errors.

In this particular case, we we're trying to help out with the common
error of omitting a scheme. With this change in place, URLs like
"www.google.com" and "teecom.com" will now work as expected.

Fixes #23